Add handleVideoStation.ts
